Move document publish/unpublish tests to own file
This keeps the test suites more focussed and easier to understand.
